chainguard/tea_integration.py
"""Mock Tea Protocol integration for ChainGuard."""
import yaml
import logging
import requests

class TeaProtocolClient:
    """
    Mock client for interacting with the Tea Protocol.

    Attributes:
        config (dict): Configuration loaded from config.yaml.
    """
    def __init__(self):
        """Initialize the Tea Protocol client with configuration."""
        try:
            with open("config.yaml", "r") as f:
                self.config = yaml.safe_load(f)
            logging.info("Initialized Tea Protocol client")
        except FileNotFoundError:
            logging.error("config.yaml not found")
            raise FileNotFoundError("Configuration file config.yaml not found")

    def get_package_metadata(self, package_name):
        """
        Fetch package metadata and teaRank from Tea Protocol.

        Args:
            package_name (str): Name of the package.

        Returns:
            dict: Package metadata including teaRank and dependency graph.
        """
        try:
            # Mock API call to Tea Protocol
            response = {
                "package": package_name,
                "teaRank": 0.85,  # Mock teaRank score
                "dependencies": ["dependency1", "dependency2"],
                "dependents": ["dependent1"]
            }
            logging.debug(f"Fetched metadata for {package_name}: {response}")
            return response
        except Exception as e:
            logging.error(f"Error fetching metadata for {package_name}: {str(e)}")
            raise

    def submit_vulnerability(self, package_name, description):
        """
        Submit a vulnerability report with TEA token staking.

        Args:
            package_name (str): Name of the package.
            description (str): Vulnerability description.

        Returns:
            str: Transaction ID or confirmation message.
        """
        try:
            # Mock Tea Protocol submission
            response = f"Submitted vulnerability for {package_name}: {description}"
            logging.debug(f"Submitted vulnerability: {response}")
            return response
        except Exception as e:
            logging.error(f"Error submitting vulnerability for {package_name}: {str(e)}")
            raise

    def manage_treasury(self):
        """
        Manage on-chain treasury for TEA token rewards.

        Returns:
            dict: Treasury status.
        """
        try:
            # Mock treasury management
            return {"treasury_address": self.config["tea"]["treasury_address"], "balance": 1000}
        except Exception as e:
            logging.error(f"Error managing treasury: {str(e)}")
            raise