requirements.txt
pip-audit==2.6.1
click==8.1.7
flask==3.0.3
requests==2.32.3
pyyaml==6.0.1
pytest==8.3.2
flask-httpauth==4.8.0pip-audit==2.6.1
